requirements.txt
anyio==3.7.0
bcrypt==4.0.1
click==8.1.3
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pydantic==1.10.8
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
